First version of refactored messaging
Completely throws away the fragile code generation and replaces it with a proper Javascript file that's inserted into the page. This script connects to the background script directly and sends its data through a Chrome runtime Port. This simplifies a lot of things greatly and makes it easier to maintain and add new features.